Add full %t/$timeformat semantics (IEEE 1800 §21.3.2) + widen proc_multipliers to u64
Fresh-probe sweep (35 probes) found %t was a broad oracle-divergent
silent-wrong: rendered as plain %0d (docs caveat only, no use-site
diagnostic) — no rescale of the module-unit time value to $timeformat
units (default = the global precision, so 1ns/1ps must show ps) and no
default min field width 20. $timeformat itself was a W3056 warn-skip.

Mechanism (all IR-0, format_version 19 unchanged):
- $timeformat lowers as a no-op SysTaskId::Display + timeformat_stmts
  StmtId side table (the severities/assert_ctl precedent — no new frozen
  SysTaskId, no format bump). The engine dispatch intercept evaluates the
  args at EXECUTION time: runtime-variable args are legal, ordering is
  last-wins, 0 args resets the defaults, 1-3 args is a loud E3009
  (iverilog: "zero or four arguments"). Hostile caps: units global±64,
  prec ≤64, |minw| ≤4096 (the WIDE-ARITH-CAP pattern).
- %t renders via fmt_time_spec: integer path = exact decimal-string math
  with TRUNCATION at the precision digit (iverilog never rounds:
  9995 @ -6/prec1 → "9.9"); real path = f64 with %f rounding (5.556 →
  "5.56"); unknown-bearing ints keep the collapsed x/X/z/Z char while the
  net shift appends zeros ("X000", "X.00 ns") and go numeric with unknown
  bits cleared once the shift cuts into digits ("0.00us"); %Nt/%0t/%0Nt
  (zero-fill before the sign) override the min width; a negative min
  width left-justifies; the suffix %s-coerces (string literal = exact
  text, every other expr = the packed-ASCII value path).
- global_prec_exp plumbed into SimOpts (already in the staged timescale
  trailer); timeformat_stmts appended as the StagedExtraSidecars tail
  (wire fixture re-pinned).

Adversarial review, two rounds:
- R1 differential caught %0Nt dropping the zero-fill flag (fixed:
  min_zero threaded into fmt_time_spec; zeros go before the minus sign,
  unlike the sign-aware %0Nd — iverilog-pinned).
- R1 soundness found a CRITICAL pre-existing companion: proc_multipliers
  was Vec<u32>, so a timescale ratio ≥10^10 (1s/1ps → M=10^12) SATURATED
  at u32::MAX (a non-power-of-10), making $time and %t off by decades.
  Widened to Vec<u64> end-to-end; postcard varint keeps old artifacts
  wire-compatible (values < 2^32 encode identically — pinned by
  timescale_trailer_u32_to_u64_wire_compat).
- R1 soundness also found $timeformat as a deferred-assert action would
  be captured by the §16.4 push_stmt hook and silently print its args at
  maturation (try_defer intercepts before the timeformat check) → loud
  E3009 reject; and unified the suffix Const/eval paths (a literal and a
  reg holding the same bits now render identically).
- R2 focused differential: 16 probes SAME, %d/%h/%b regression-free,
  staged (vcmp→velab→vrun) byte-equal to one-shot at 1s/1ps and 100s/1fs.

Byte-identity: designs without %t and $timeformat are untouched (empty
side table, %t arm only path change); %0t with M=1 and no $timeformat
renders exactly the old fmt_dec output. The only expectation update is
uppercase_format.rs (bare %t now width-20, the oracle-correct form).

Accepted documented divergence: negative time value scale-DOWN — iverilog
13.0 prints corrupt output ("0.0-u", self-inconsistent), vita renders the
sign-aware "-0.00u".

Docs: hdl-reference 01/05 %t caveat resolved; ROADMAP §4.5.60 (+ 9
adjacent probe-harvest candidates recorded: pre-opened FDs, %- flag
family, queue slice, ==?, force part-select, hierarchical func call,
$printtimescale, %t-of-string cross-spec, sigpipe test flake); CLAUDE.md
status blocks compressed per the history-management policy.

21 new tests (crates/cli/tests/timeformat.rs, all live iverilog byte
pins + staged parity + arity-loud + defer-loud). 2746 tests green,
clippy/fmt clean.

+++ 016_claude_rtl/crates/sim-engine/src/sched.rs
@@ -1026,7 +1026,7 @@ impl<'a, 'ir> Scheduler<'a, 'ir> {
             .get(tmpl)
             .copied()
             .unwrap_or(1)
-            .max(1) as u64;
+            .max(1);
         // VM-REGPOOL: lease the register/offset files from the pool, sized to this
         // body, and return them afterwards (a `pop` yields an OWNED buffer, so it no
         // longer borrows `self` and cannot alias the `&mut self` kernel call).
